:sparkles: member api 정의
