+++ src/apis/schema/record.ts
@@ -15,11 +15,3 @@ export interface RecordType {
   startedAt: string;
   finishedAt: string;
 }
-
-export type ImageFileExtensionType = 'JPG' | 'JPEG' | 'PNG';
-
-export const IMAGE_File_Extension: Record<string, ImageFileExtensionType> = {
-  'image/jpeg': 'JPEG',
-  'image/png': 'PNG',
-  'image/jpg': 'JPG',
-};
